Added some options to the Play menu:  * set A marker  * set B marker  * clear A-B markers
This allows to select an A-B section in the file. Only that section will be
played. If the option Play->Repeat is enabled, the A-B section will be 
repeated (if there's no A-B section, Play->Repeat will repeat the whole file).

Notice: when there's an A-B section, any action which requires to restart
mplayer, will start the playback at the A marker.